Bluetooth: roll back phonebook sync changes
well it fixed for a few carcits the sync but broke it for way
more after going trough all the reports. For now due that we go
into stable direction roll back and we need to check deeper where the
problem is. AOSP did not push this changes trough till now as well

Change-Id: I35fcd6f96359ef1310bfc3c12d9c2859b5e5c301